finishing trydjango 1.9 basic

+++ src/posts/views.py
@@ -1,11 +1,22 @@
 from django.shortcuts import render,get_object_or_404,reverse,redirect
 from django.core.paginator import Paginator,EmptyPage,PageNotAnInteger
+from django.db.models import Q
+from urllib.parse import quote_plus
 from django.contrib import messages
 from .models import Post
-from .forms import PostForm
+from .forms import PostForm,SearchForm
 # Create your views here.
 def posts_home(request):
-    queryset_list=list(Post.objects.all())
+    queryset_list=Post.objects.all()
+    query=request.GET.get('query') or request.POST.get('query') or None
+    if query:
+        print(query)
+        queryset_list=queryset_list.filter(
+            Q(title__icontains=query)|
+            Q(content__icontains=query)|
+            Q(user__first_name__icontains=query)|
+            Q(user__last_name__icontains=query)
+        ).distinct()
     paginator=Paginator(queryset_list,4)
     page=request.GET.get('page')
     try:
@@ -18,6 +29,8 @@ def posts_home(request):
     contex={
         'name':'seam',
         'obj_list':queryset,
+        'searched_value':query,
+        'search_form':SearchForm,
     }
     return render(request,'posts/index.html',contex)
 
@@ -33,16 +46,19 @@ def posts_create(request):
     }
     return render(request,'posts/form.html',context)
 
-def posts_delete(request,post_id=None):
-    obj=get_object_or_404(Post,id=post_id)
-    if request.method=="POST":
-        obj.delete()
-        return redirect(reverse('posts:posts_home'))
+def posts_delete(request,slug=None):
+    obj=get_object_or_404(Post,slug=slug)
+    if request.user==obj.user:
+        if request.method=="POST":
+            obj.delete()
+            return redirect(reverse('posts:posts_home'))
+    else:
+        raise Http403
     return render(request,'posts/delete.html',{"obj":obj})
 
 
-def posts_update(request,post_id=None):
-    obj=get_object_or_404(Post,id=post_id)
+def posts_update(request,slug=None):
+    obj=get_object_or_404(Post,slug=slug)
     if request.method=="POST":
         form=PostForm(request.POST,request.FILES,instance=obj)
         if form.is_valid():
@@ -60,6 +76,11 @@ def posts_update(request,post_id=None):
 
     return render(request,'posts/form.html',context)
 
-def posts_detail(request,post_id=None):
-    obj=get_object_or_404(Post,id=post_id)
-    return render(request,'posts/detail.html',{'obj':obj})
+def posts_detail(request,slug=None):
+    obj=get_object_or_404(Post,slug=slug)
+    print(request.build_absolute_uri)
+    context={
+        'obj':obj,
+        'share_string':quote_plus(obj.content)
+    }
+    return render(request,'posts/detail.html',context)
